+++ R/az_kubernetes_service.R
@@ -7,6 +7,10 @@
 #' The following methods are available, in addition to those provided by the [AzureRMR::az_resource] class:
 #' - `new(...)`: Initialize a new AKS object.
 #' - `get_cluster(config, role)`: Return an object representing the Docker registry endpoint.
+#' - `get_agent_pool(pollname)`: Returns an object of class `az_agent_pool` representing an agent pool. This class inherits from [AzureRMR::az_resource]; it currently has no extra methods, but these may be added in the future.
+#' - `list_agent_pools()`: Returns a list of agent pool objects.
+#' - `create_agent_pool(poolname, ..., wait=FALSE)`: Creates a new agent pool. See the [agent_pool] function for further arguments to this method.
+#' - `delete_agent_pool(poolname, confirm=TRUE. wait=FALSE)`: Deletes an agent pool.
 #' - `list_cluster_resources()`: Returns a list of all the Azure resources managed by the cluster.
 #' - `update_aad_password(name=NULL, duration=NULL, ...)`: Update the password for Azure Active Directory integration, returning the new password invisibly. See 'Updating credentials' below.
 #' - `update_service_password(name=NULL, duration=NULL, ...)`: Update the password for the service principal used to manage the cluster resources, returning the new password invisibly.  See 'Updating credentials' below.
@@ -22,9 +26,9 @@
 #' - `role`: This can be `"User"` (the default) or `"Admin"`.
 #'
 #' @section Updating credentials:
-#' An AKS cluster requires at least one, and possibly three, service principals. The first service principal is used to manage the resources used by the cluster: the VMs, networking resources, virtual disks, etc. The other two are used for AAD integration. These service principals have secret passwords, which have to be refreshed as they expire.
+#' An AKS resource can have up to three service principals associated with it. Two of these are for Azure Active Directory (AAD) integration. The third is used to manage the subsidiary resources (VMs, networks, disks, etc) used by the cluster, if it doesn't have a service identity.
 #'
-#' The `update_aad_password()` and `update_service_password()` methods let you refresh the passwords for the cluster's service principals. Their arguments are:
+#' Any service principals used by the AKS resource will have secret passwords, which have to be refreshed as they expire. The `update_aad_password()` and `update_service_password()` methods let you refresh the passwords for the cluster's service principals. Their arguments are:
 #'
 #' - `name`: An optional friendly name for the password.
 #' - `duration`: The duration for which the new password is valid. Defaults to 2 years.
@@ -53,7 +57,14 @@
 #' # get the cluster endpoint
 #' kubclus <- myaks$get_cluster()
 #'
-#' # refresh the service principal password
+#' # list of agent pools
+#' myaks$list_agent_pools()
+#'
+#' # create a new agent pool, then delete it
+#' pool <- myaks$create_agent_pool("pool2", 3, size="Standard_DS3_v2")
+#' pool$delete()
+#'
+#' # refresh the service principal password (mostly for legacy clusters without a managed identity)
 #' myaks$update_service_password()
 #'
 #' # refresh the service principal password, using custom credentials to authenticate with MS Graph
@@ -104,6 +115,43 @@ public=list(
         kubernetes_cluster(config=config)
     },
 
+    get_agent_pool=function(poolname)
+    {
+        az_agent_pool$new(self$token, self$subscription, resource_group=self$resource_group,
+            type="Microsoft.ContainerService/managedClusters",
+            name=file.path(self$name, "agentPools", poolname),
+            api_version=self$get_api_version())
+    },
+
+    create_agent_pool=function(poolname, ..., wait=FALSE)
+    {
+        pool <- agent_pool(name=poolname, ...)
+        az_agent_pool$new(self$token, self$subscription, resource_group=self$resource_group,
+            type="Microsoft.ContainerService/managedClusters",
+            name=file.path(self$name, "agentPools", poolname),
+            properties=pool[-1],
+            api_version=self$get_api_version(),
+            wait=wait)
+    },
+
+    delete_agent_pool=function(poolname, confirm=TRUE, wait=FALSE)
+    {
+        az_agent_pool$new(self$token, self$subscription, resource_group=self$resource_group,
+            type="Microsoft.ContainerService/managedClusters",
+            name=file.path(self$name, "agentPools", poolname),
+            deployed_properties=list(NULL),
+            api_version=self$get_api_version())$
+            delete(confirm=confirm, wait=wait)
+    },
+
+    list_agent_pools=function()
+    {
+        cont <- self$do_operation("agentPools")
+        api_version <- self$get_api_version()
+        lapply(get_paged_list(cont, self$token), function(parms)
+            az_agent_pool$new(self$token, self$subscription, deployed_properties=parms, api_version=api_version))
+    },
+
     list_cluster_resources=function()
     {
         clusrgname <- self$properties$nodeResourceGroup
@@ -129,6 +177,9 @@ public=list(
     update_service_password=function(name=NULL, duration=NULL, ...)
     {
         prof <- self$properties$servicePrincipalProfile
+        if(prof$clientId == "msi")
+            stop("Cluster uses service identity for managing resources", call.=FALSE)
+
         app <- graph_login(self$token$tenant, ...)$get_app(prof$clientId)
         app$add_password(name, duration)
         prof$secret <- app$password
@@ -146,15 +197,86 @@ public=list(
 #' @param count The number of nodes per pool.
 #' @param size The VM type (size) to use for the pool. To see a list of available VM sizes, use the [list_vm_sizes] method for the resource group or subscription classes.
 #' @param os The operating system to use for the pool. Can be "Linux" or "Windows".
+#' @param disksize The OS disk size in gigabytes for each node in the pool. A value of 0 means to use the default disk size for the VM type.
+#' @param use_scaleset Whether to use a VM scaleset instead of individual VMs for this pool. A scaleset offers greater flexibility than individual VMs, and is the recommended method of creating an agent pool.
+#' @param low_priority If this pool uses a scaleset, whether it should be made up of spot (low-priority) VMs. A spot VM pool is cheaper, but is subject to being evicted to make room for other, higher-priority workloads. Ignored if `use_scaleset=FALSE`.
+#' @param cluster_autoscale The autoscaling parameters for the pool. This can be either `FALSE`, meaning autoscaling is disabled, or a vector of 2 numbers giving the minimum and maximum size of the agent pool. Ignored if `use_scaleset=FALSE`.
+#' @param ... Other named arguments, to be used as parameters for the agent pool.
+#'
+#' @details
+#' `agent_pool` is a convenience function to simplify the task of specifying the agent pool for a Kubernetes cluster.
+#'
+#' @return
+#' An object of class `agent_pool`, suitable for passing to the `create_aks` constructor method.
+#'
+#' @seealso
+#' [create_aks], [list_vm_sizes]
+#'
+#' [Agent pool parameters on Microsoft Docs](https://docs.microsoft.com/en-us/rest/api/aks/managedclusters/createorupdate#managedclusteragentpoolprofile)
+#'
+#' @examples
+#' # pool of 5 Linux GPU-enabled VMs
+#' agent_pool("pool1", 5, size="Standard_NC6s_v3")
+#'
+#' # pool of 3 Windows Server VMs, 500GB disk size each
+#' agent_pool("pool1", 3, os="Windows", disksize=500)
+#'
+#' # enable cluster autoscaling, with a minimum of 1 and maximum of 10 nodes
+#' agent_pool("pool1", 5, cluster_autoscale=c(1, 10))
+#'
+#' # use individual VMs rather than scaleset
+#' agent_pool("vmpool1", 3, use_scaleset=FALSE)
+#'
+#' @export
+agent_pool <- function(name, count, size="Standard_DS2_v2", os="Linux", disksize=0,
+                       use_scaleset=TRUE, low_priority=FALSE, cluster_autoscale=FALSE, ...)
+{
+    parms <- list(
+        name=name,
+        count=as.integer(count),
+        vmSize=size,
+        osType=os,
+        osDiskSizeGB=disksize,
+        type=if(use_scaleset) "VirtualMachineScaleSets" else "AvailabilitySet"
+    )
+
+    if(use_scaleset)
+    {
+        if(is.numeric(cluster_autoscale) && length(cluster_autoscale) == 2)
+        {
+            parms$enableAutoScaling <- TRUE
+            parms$minCount <- min(cluster_autoscale)
+            parms$maxCount <- max(cluster_autoscale)
+        }
+        if(low_priority)
+            parms$scaleSetPriority <- "spot"
+    }
+
+    extras <- list(...)
+    if(!is_empty(extras))
+        parms <- utils::modifyList(parms, extras)
+
+    structure(parms, class="agent_pool")
+}
+
+
+#' Vectorised utility function for specifying Kubernetes agent pools
+#'
+#' @param name The name(s) of the pool(s).
+#' @param count The number of nodes per pool.
+#' @param size The VM type (size) to use for the pool. To see a list of available VM sizes, use the [list_vm_sizes] method for the resource group or subscription classes.
+#' @param os The operating system to use for the pool. Can be "Linux" or "Windows".
 #'
 #' @details
 #' This is a convenience function to simplify the task of specifying the agent pool for a Kubernetes cluster. You can specify multiple pools by providing vectors as input arguments; any scalar inputs will be replicated to match.
+#'
+#' `aks_pools` is deprecated; please use [agent_pool] going forward.
 #'
 #' @return
 #' A list of lists, suitable for passing to the `create_aks` constructor method.
 #'
 #' @seealso
-#' [list_vm_sizes]
+#' [list_vm_sizes], [agent_pool]
 #'
 #' @examples
 #' # 1 pool of 5 Linux VMs
@@ -174,4 +296,3 @@ aks_pools <- function(name, count, size="Standard_DS2_v2", os="Linux")
     pool_df$name <- make.unique(pool_df$name, sep="")
     lapply(seq_len(nrow(pool_df)), function(i) unclass(pool_df[i, ]))
 }
-