+++ tests/testthat/test02_aci.R
@@ -8,19 +8,33 @@ subscription <- Sys.getenv("AZ_TEST_SUBSCRIPTION")
 if(tenant == "" || app == "" || password == "" || subscription == "")
     skip("Tests skipped: ARM credentials not set")
 
-acrname <- Sys.getenv("AZ_TEST_ACR")
-if(acrname == "")
-    skip("ACI tests skipped: resource names not set")
+rgname <- make_name(10)
+rg <- AzureRMR::az_rm$
+    new(tenant=tenant, app=app, password=password)$
+    get_subscription(subscription)$
+    create_resource_group(rgname, location="australiaeast")
+
+echo <- getOption("azure_containers_tool_echo")
+options(azure_containers_tool_echo=FALSE)
 
 test_that("ACI works",
 {
-    rgname <- Sys.getenv("AZ_TEST_RG")
-    rg <- AzureRMR::az_rm$
-        new(tenant=tenant, app=app, password=password)$
-        get_subscription(subscription)$
-        get_resource_group(rgname)
+    acrname <- make_name(10)
+    acr <- rg$create_acr(acrname, admin_user_enabled=TRUE)
+    reg <- acr$get_docker_registry(as_admin=TRUE)
+    expect_true(is_docker_registry(reg))
+    expect_false(is.null(reg$username) || is.null(reg$password))
 
-    aciname <- paste0(sample(letters, 10, TRUE), collapse="")
+    cmdline <- "build -f ../resources/hello_dockerfile -t hello-world ."
+    call_docker(cmdline)
+
+    reg$push("hello-world")
+
+    cmdline <- paste0("image rm ", acrname, ".azurecr.io/hello-world")
+    call_docker(cmdline)
+
+    # from local image
+    aciname <- make_name(10)
     expect_true(is_aci(rg$create_aci(aciname,
         image="hello-world")))
 
@@ -32,23 +46,25 @@ test_that("ACI works",
     expect_silent(aci$start())
     expect_silent(aci$restart())
 
-    acr <- rg$get_acr(acrname)
-    expect_true(is_acr(acr))
-    reg <- acr$get_docker_registry(as_admin=TRUE)
-    expect_true(is_docker_registry(reg))
-    expect_false(is.null(reg$username) || is.null(reg$password))
-
-    aciname2 <- paste0(sample(letters, 10, TRUE), collapse="")
+    # from Resource Manager object
+    aciname2 <- make_name(10)
     aci2 <- rg$create_aci(aciname2,
         image=paste0(reg$server$hostname, "/hello-world"),
         registry_creds=reg)
 
     expect_true(is_aci(aci2))
 
-    aciname3 <- paste0(sample(letters, 10, TRUE), collapse="")
+    # from Docker registry object
+    aciname3 <- make_name(10)
     aci3 <- rg$create_aci(aciname3,
-        image=paste0(reg$server$hostname, "/hello-world-sp"),
+        image=paste0(reg$server$hostname, "/hello-world"),
         registry_creds=aci_creds(reg$server$hostname, app, password))
 
     expect_true(is_aci(aci3))
+})
+
+
+teardown({
+    options(azure_containers_tool_echo=echo)
+    suppressMessages(rg$delete(confirm=FALSE))
 })